Complete AST module refactoring (DML, SELECT, statement)
Extracted remaining statement types from lib.rs into focused modules:
- dml.rs (39 lines): INSERT, UPDATE, DELETE statements
- select.rs (70 lines): SELECT statement and related types
- statement.rs (20 lines): Top-level Statement enum

Refactored lib.rs to become thin re-export layer (19 lines of exports).

All 22 AST tests passing, no clippy warnings.
Files appropriately sized and well-organized by responsibility.

Closes #31
