+++ crates/ast/src/statement.rs
@@ -1,0 +1,20 @@
+//! Top-level SQL statement types
+//!
+//! This module defines the Statement enum that represents all possible SQL statements.
+
+use crate::{CreateTableStmt, DeleteStmt, InsertStmt, SelectStmt, UpdateStmt};
+
+// ============================================================================
+// Top-level SQL Statements
+// ============================================================================
+
+/// A complete SQL statement
+#[derive(Debug, Clone, PartialEq)]
+pub enum Statement {
+    Select(SelectStmt),
+    Insert(InsertStmt),
+    Update(UpdateStmt),
+    Delete(DeleteStmt),
+    CreateTable(CreateTableStmt),
+    // TODO: Add more statement types (ALTER, DROP, etc.)
+}
